flatpak: Handle flatpak+https URI scheme for flatpakrefs
Since web browsers no longer make it easy to open a downloaded file with
a specific (or even the default) application as soon as it’s downloaded
– they now just download to disk – flathub wants to use a different URI
scheme to link to flatpakref files, so that they will get loaded
directly in gnome-software or KDE Discover without the user having to
find them in the Downloads folder and launch them.

Implement support for that, if the flatpak plugin is enabled.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #2240